// web/src/api/generated/NextPayeeDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NodeId } from "./NodeId";
import type { Tier } from "./Tier";

/**
 * Predicted payee of the next block for one tier.
 */
export type NextPayeeDto = { tier: Tier, node: NodeId | null, address: string, };